==> 1. data/1. fruit_fly_data/1. obtain_neuron_info_and_skeletons.ipynb <==
# Whole skeleton neuron df queying

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
from neuprint import fetch_synapses, NeuronCriteria as NC, SynapseCriteria as SC
from neuprint import Client
from neuprint import fetch_adjacencies, NeuronCriteria as NC
from neuprint import fetch_neurons
from neuprint import merge_neuron_properties
import math
import json
from neuprint import fetch_synapse_connections
from neuprint import Client
from urllib.error import URLError, HTTPError
from urllib.request import urlopen
import timeit
import pickle

import numpy as np
import pandas as pd
from scipy.spatial import KDTree
import matplotlib.pyplot as plt
import pickle
import timeit
import itertools
from neuprint import fetch_roi_hierarchy
import networkx as nx
from scipy.spatial import distance

TOKEN = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Imx1a2FibGFnb2pldmljMTk5NUBnbWFpbC5jb20iLCJsZXZlbCI6Im5vYXV0aCIsImltYWdlLXVybCI6Imh0dHBzOi8vbGgzLmdvb2dsZXVzZXJjb250ZW50LmNvbS9hLS9BT2gxNEdqdDZpdFFGR2xTSWZUTElNTjRmcEt1QzZ3QmE2Rlp0WU1XYmpKV1ZBPXM5Ni1jP3N6PTUwP3N6PTUwIiwiZXhwIjoxODA5MTE1OTEyfQ.0h6CJp8xfQEpkW8a2_gqJUBrEA5GyBiZkNvDjRpoXoY" # <--- Paste your token here
           # (or define NEUPRINT_APPLICATION CREDENTIALS in your environment)

c = Client('neuprint.janelia.org', 'hemibrain:v1.2.1', TOKEN)

dataset_lengths = {}
path_save_1 = '0. neuron_regions_df'
path_save_2 = '1. neuron_regions_skeleton'
neuron_df_dict = {}
minimum_synapses = 10
regions = ['GF(R)','GC','mALT(L)','POC']
for key in regions:
    print(key)
    if key[-1:] == '*':
        key = key[:-1]
    else:
        key = key
    criteria = NC(min_roi_inputs=minimum_synapses,min_roi_outputs=minimum_synapses, inputRois=[key], outputRois=[key])
    neuron_df, roi_counts_df = fetch_neurons(criteria)
    dataset_lengths[key] = len(neuron_df)
    neuron_df.to_csv(path_save_1 + key+ '.csv')
    neuron_df_dict[key] = neuron_df 

neuron_df.head()

columns = ['x','y','z','radius','bodyId']
df_dict = {}
skeleton_df = pd.DataFrame(columns=columns)
for neuron_df_key in regions:
    print(neuron_df_key)
    neuron_df = neuron_df_dict[neuron_df_key]
    neuron_list = neuron_df['bodyId'].to_list()
    for i,bodyId in enumerate(neuron_list):
        try:
            s =  c.fetch_skeleton(body = int(bodyId),format='pandas')
            s['bodyId'] = str(bodyId)
            skeleton_df = pd.concat([skeleton_df,s])
            count += 1
            count_save +=1
            
        except:
            pass
            #print("Neuron was not loaded")
    skeleton_df = skeleton_df[['x','y','z','radius','bodyId']]
    skeleton_df.to_csv(path_save_2 + neuron_df_key+'_'+str(len(neuron_df)) + '.csv')
    df_dict[neuron_df_key] = skeleton_df

skeleton_df.head()
